Creating Member Vars
Next we need to create copy contructors and overload assignment ops for all objects. This way we can dynamically change the size of a UMLClass depending on the lenght of the member variable name. After that, we should be able to create a whole application that creates UMLClass and dynamically changes the size of the image.